Feature/like function

+++ src/main/java/com/sparta/fifteen/service/LikeService.java
@@ -1,12 +1,10 @@
 package com.sparta.fifteen.service;
 
 import com.sparta.fifteen.entity.*;
-import com.sparta.fifteen.error.CommentNotFoundException;
 import com.sparta.fifteen.error.PostNotFoundException;
 import com.sparta.fifteen.error.SelfPostLikeException;
 import com.sparta.fifteen.repository.*;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Objects;
 import java.util.Optional;
@@ -27,7 +25,6 @@ public class LikeService {
         this.newsFeedRepository = newsFeedRepository;
     }
 
-    @Transactional
     public void likeOrUnlike(User user, Long contentId, ContentTypeEnum contentType) {
 
         Long userId = user.getId();
@@ -42,24 +39,25 @@ public class LikeService {
             if(newsFeed.isEmpty()) {
                 throw new PostNotFoundException("선택된 뉴스피드는 존재하지 않습니다."); // 예외 처리로 수정
             }
-            Optional<LikeNewsFeed> existingLike = likeNewsFeedRepository.findByUserIdAndNewsfeedId(userId, contentId);
+            Optional<LikeNewsFeed> existingLike = likeNewsFeedRepository.findByNewsfeedIdAndUserId(userId, contentId);
             if (existingLike.isPresent()) {
                 likeNewsFeedRepository.delete(existingLike.get());
             } else {
-                LikeNewsFeed like = new LikeNewsFeed( user,  newsFeed.get());
+                LikeNewsFeed like = new LikeNewsFeed( userId,  contentId);
                 likeNewsFeedRepository.save(like);
             }
         }else{
-            Comment comment = commentRepository.findById(contentId).orElseThrow(() -> new CommentNotFoundException("선택된 댓글은 존재하지 않습니다."));
+            Optional<Comment> comment = commentRepository.findById(contentId);
+            if(comment.isEmpty()) {
+                throw new PostNotFoundException("선택된 댓글은 존재하지 않습니다."); // 예외 처리로 수정
+            }
 
             Optional<LikeComment> existingLike = likeCommentRepository.findByUserIdAndCommentId(userId, contentId);
             if (existingLike.isPresent()) {
                 likeCommentRepository.delete(existingLike.get());
-                comment.updatelikes(-1L);
             } else {
-                LikeComment like = new LikeComment( user,  comment);
+                LikeComment like = new LikeComment( userId,  contentId);
                 likeCommentRepository.save(like);
-                comment.updatelikes(1L);
             }
         }
